requirements.txt
streamlit==1.37.1
duckdb>=1.0.0
pandas>=2.0.0
altair>=5.0.0
openai>=1.40.0
numpy>=1.24.0
python-dotenv
